+++ CHANGELOG.md
@@ -1,8 +1,28 @@
-## 0.4.0
+## [0.4.6] - 2026-01-24
+- fix: numeric keyboard not hiding on un-focus
+- fix: apple app links support
+
+## [0.4.5] - 2025-12-15
+- add relay connection indicator in app bar 
+- improve relay connection handling
+- connect wallet popup with Alby Go 1-click connection
+- NWC connection QR scanner screen
+
+
+## [0.4.4] - 2025-12-13
+### Fixed
+- fix reconnect to relays when app goes to background
+
+## [0.4.3]
+- 🇮🇹 italian translation
+- improve FAQ
+- don't check coordinator health immediately, assume healthy
+
+## [0.4.0]
 - initial NWC wallet support
 - fix offer's list state update bug
 
-## 0.3.0
+## [0.3.0]
 
 - use nostr relays & websockets for client/coordinator communication
 - backup/restore private keys